Have image format control iteration order
To prepare for incremental image output, allow the format to dictate the
correct order for scanlines of the image to be rendered.  This allows
the format to cause the renderer to process scanlines in the order that
allows the lines to be written as they would appear in the file format
on disk.